소린 helps 상록 item_cart, using cart_provider method

+++ lib/provider/cart_provider.dart
@@ -44,7 +44,7 @@ class CartProvider with ChangeNotifier {
     }
   }
 
-  double get totalPrice =>
+  int get totalPrice =>
       _cartItems.fold(0, (sum, e) => sum + (e.item.price * e.quantity));
 
   void clearCart() {
